requirements.txt
# Core
apache-airflow==2.9.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pyarrow==15.0.2

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31

# API REST
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Modelos e MLflow
mlflow==2.14.3
lightgbm==4.3.0
xgboost==2.0.3
flaml[automl]==2.1.2
torch==2.10.0
joblib==1.4.2
